Adds the ability to treat property values as URIs when mapping to RDF
The initial implementation assumed that all property value pairs on
vertices would be mapped such that the property is used to create a
predicate URI and the value is used to create a literal.

However in some cases (e.g. adding type descriptors as seen in the
example script) you actually want to have the value of some properties
treated as URIs.  This commit adds a new uriProperties field to the RDF
mapping which allows users to specify which property values should be
mapped as such.